refactor: move ChoicePicker utils to core/include
- Hoisting choicepicker_component_utils.h into core/include/ restores
  the one-way dependency rule and lets all three platform renderers
  share the same parsing helpers.
