+++ platforms/harmony/agenui/src/main/cpp/a2ui/render/components/choicepicker_component.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "../a2ui_component.h"
-#include "choicepicker_component_utils.h"
+#include "agenui_choicepicker_component_utils.h"
 #include <vector>
 
 namespace a2ui {
